Refactor database migrations and update model snapshots
- Deleted obsolete migration files for users and orders.
- Created new migration files for PostgreSQL with updated data types and constraints.
- Updated model snapshot to reflect the latest database schema.
- Adjusted product prices in the migration data to use consistent decimal formatting.
- Updated assembly information for the project.

+++ WebShopBackend/Data/ShopContext.cs
@@ -28,10 +28,10 @@ public class ShopContext : DbContext
             .WithMany()
             .HasForeignKey(oi => oi.ProductId);
         modelBuilder.Entity<Product>().HasData(
-            new Product { Id = 1, Name = "Premium Olivenöl",     Price = 18.50m, Category = "Öl",    ImageUrl = "wine01.jpeg" },
-            new Product { Id = 2, Name = "Rotwein Riserva",      Price = 24.90m, Category = "Wein",  ImageUrl = "wine01.jpeg" },
-            new Product { Id = 3, Name = "Weißwein Chardonnay",  Price = 14.20m, Category = "Wein",  ImageUrl = "wine01.jpeg" },
-            new Product { Id = 4, Name = "Balsamico Essig",      Price = 12.00m, Category = "Essig", ImageUrl = "wine01.jpeg" }
+            new Product { Id = 1, Name = "Premium Olivenöl",    Price = (decimal)18.50, Category = "Öl",    ImageUrl = "wine01.jpeg" },
+            new Product { Id = 2, Name = "Rotwein Riserva",     Price = (decimal)24.90, Category = "Wein",  ImageUrl = "wine01.jpeg" },
+            new Product { Id = 3, Name = "Weißwein Chardonnay", Price = (decimal)14.20, Category = "Wein",  ImageUrl = "wine01.jpeg" },
+            new Product { Id = 4, Name = "Balsamico Essig",     Price = (decimal)12.00, Category = "Essig", ImageUrl = "wine01.jpeg" }
         );
     }
 }
